update test environments
