+++ Makefile
@@ -1,5 +1,5 @@
 CI_BUILD_NUMBER ?= $(USER)-snapshot
-VERSION ?= 17.1.$(CI_BUILD_NUMBER)
+VERSION ?= 17.2.$(CI_BUILD_NUMBER)
 
 version:
 	@echo $(VERSION)
