+++ packages/mwp-tracking-plugin/README.md
@@ -42,7 +42,7 @@ Note that we do not currently have a unique place to store tracking data from
 dev, so any tracking data that you produce will be merged into production
 tracking data. For small amounts of data, this shouldn't significantly affect
 analytics.
- 
+
 ## Activity tracking
 
 Activity tracking is provided by a `request.trackActivity` method, which has two
@@ -53,12 +53,13 @@ responsibilities:
 
 ## Click tracking
 
-Click tracking consists of 3 related modules:
+Click tracking consists of 4 related modules:
 
 1. `clickWriter` for writing the click `Event` data to a Redux action.
 2. `clickState` for defining Redux click actions and reducing those actions into
-   Redux state, and writing click tracking data into a cookie.
-3. `clickReader` for reading the click tracking data cookie on the server.
+   a click history object, and writing click tracking data into a cookie.
+3. `clickMiddleware` for listening for click actions and updating the click cookie
+4. `clickReader` for reading the click tracking data cookie on the server.
 
 Ultimately, the only 'fixed' requirements are that the click tracking data must
 be passed to the server in a cookie, and the serialization of that data must be
@@ -68,23 +69,35 @@ with parallel behavior implemented in Meetup Classic.
 ### Behavior
 
 The platform currently adds `click` _and_ `change` listeners when creating the
-Redux store for the browser - this is done on application setup and is
+Redux store for the browser - this is done on application start and is
 guaranteed not to run on the server. Each `click` and `change` event is sent
 to `clickWriter`, which dispatches a Redux `CLICK_TRACK` action.
 
-The click reducer is built into the platform reducer, and will store
-click tracking data for the lifetime of the session, or until it is cleared with
-a particular click 'clear' Redux action.
+The click action is _not_ reduced into Redux state, but is picked up by middleware
+in order to accumulate the data in a cookie that will collect data from multiple
+tabs and persist when navigating away from the web app.
 
-When the user navigates to a new page, the generated API request will read the
-click tracking data from state, clear the click tracking state, and then set the
-click tracking cookie. The cookie value will be sent with the API request, read
-by the server, and then the server will force-clear the cookie.
+Server routes must opt-in to consuming the click cookie using `route.options.plugins['mwp-tracking-plugin']`,
+which should define an object with a `click` property defining a function that
+accepts the `request` as an argument and returns `true` when the click cookie
+should be read (e.g. based on querystring parameters or other header info).
 
-### Leakage
+Example:
 
-Because click data only gets sent to the server when a user navigates to another
-location within the application, there is a possibility of losing data when a
-user navigates to another site, or even back to Meetup Classic, because a
-'location change' action will not be dispatched and the Redux store will be
-destroyed upon navigation.
+```js
+const route = {
+	method: 'GET',
+	path: '/foo',
+	config: {
+		plugins: {
+			'mwp-tracking-plugin': {
+				// only track clicks when `?trackClicks=true` in querystring
+				click: request => Boolean(request.query.trackClicks),
+			},
+		},
+	},
+};
+```
+
+The click plugin will 'unset' the click tracking cookie when the data has been
+consumed.
